wip(weather): added get weather threshold & url builder
